IMPALA-10937: Fix broken-build on Centos-7
A recent patch upgraded thrift_sasl from 0.4.2 to 0.4.3. It broke
the builds on Centos-7.
The version of setuptools which is installed by Jenkins for the
virtualenvs on Centos 7 could be lower than the requirement by
thrift_sasl 0.4.3. Another issue is the new string syntax for
install_requires in setup.py is not accepted on Centos-7.
This patch remove the requirement for setuptools in setup.py of
thrift_sasl as work around. Also change back the syntax of strings
for install_requires.

Testing:
  - Ran core test on impala-private-parameterized which use Centos-7.
  - Ran core test on pre-review-test which use Ubuntu.

Change-Id: I2c256a8ec9a151bca8b3370bfce6ecebf060bad0
Reviewed-on: http://gerrit.cloudera.org:8080/17886
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>