IMPALA-10959: Reload MV as ACID tables
We observed that the event processor is broken after receiving a
partition event for materialized views (MV). This is because we are
treating MV as view in Impala but Hive generates partition events for MV,
which breaks current event processor.

In this patch, we let partition events of MV follow the code path of ACID
tables to reload the view. In the long term, we will need IMPALA-10723 to
treat materialized view as a table.

Tests:
- manually testing

Change-Id: Ibeab8cc53ad47d24df8baba81e1ec6ea4c80a084
Reviewed-on: http://gerrit.cloudera.org:8080/17911
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Vihang Karajgaonkar <vihang@cloudera.com>