IMPALA-10975: Refactor alter table ddl operation
For almost all alter table DDL operations in catalogOpExecutor,
a table is added to catalog update at a common place in the end
if reloadMetadata is true. However for few sub ddl operations like ADD
and DROP partitions, addTableToCatalogUpdate() is called locally.
This patch is to refactor addTableToCatalogUpdate() and call it at one
place for all the sub ddls.

After this patch, a table (after alter table ddl) is added to catalog
update if its old catalog version does not match the current catalog
version.

This patch also addresses a bug in which a HMS event was getting
added to delete event log even if event processing is not active.

Testing:
Relying on existing tests since it is a minor refactoring.

Change-Id: Ifbadcab68b4599ad18b681b1284052a47b74d802
Reviewed-on: http://gerrit.cloudera.org:8080/17919
Reviewed-by: Vihang Karajgaonkar <vihang@cloudera.com>
Tested-by: Vihang Karajgaonkar <vihang@cloudera.com>