IMPALA-10972: Fixup test behavior to work with HIVE-24920
With HIVE-24920, the HMS runs in a mode that prohibits
creating a managed directory where the default location
already exists. Some Impala test helper functions copied
files into a particular location and then created a
table without specifying the location in the create statement.
This is no longer possible.

This changes the helper functions in test/common/file_utils.py
to create the table before pulling files in.

Tests:
 - Ran a core job against a Hive with HIVE-24920 and
   verified that the failures due to changes in
   behavior are gone.
 - Ran a core job against the current Hive

Change-Id: Idfe5468a0b9e1025ec7a0ad3cdce4793f35ca7ba
Reviewed-on: http://gerrit.cloudera.org:8080/17956
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Vihang Karajgaonkar <vihang@cloudera.com>