IMPALA-10212. Support ofs scheme.
OFS is the new file system implementation for Ozone.
The biggest difference compared to o3fs is that ofs supports operations across all
volumes and buckets and provides a full view of all the volume/buckets.

It uses the same transport as o3fs and therefore it shares the thread pool with o3fs.

How it was tested:
The patch was tested manually on a CDPD cluster, loaded TPC-DS data, ran TPC-DS, ran 'load data inpath' command.

Change-Id: I69908f65c97f40ff01b25d6d6db53c37a9e978ba
Reviewed-on: http://gerrit.cloudera.org:8080/17963
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>