IMPALA-11007: Replies to HTTP HEAD requests should contain no content.
If a HEAD request to Impala's webserver, for example
  curl -I http://localhost:25000/metrics
does send content in the response then this results in ugly messages
being logged on the server side when the client side closes the
connection after it has read the headers in the http response. Fix this
by not sending (or generating) the content in reply to the HEAD
request.

TESTING:
  Ran all end-to-end tests.
  Added HEAD calls to various existing tests.

Change-Id: I5cc3bf72067dfa5bc24e29fe974e5762be507952
Reviewed-on: http://gerrit.cloudera.org:8080/18002
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>