IMPALA-11022: Impala uses wrong file descriptors for Iceberg tables in local catalog mode
When local catalog mode is used, Impala retrieves the Iceberg
snapshot from CatalogD. The response contains a map of the file
descriptors. The file descriptors contain block location information,
but the hosts are only referred to by indexes. In the Coordinator's
local catalog the host indexes might refer to different hosts than in
CatalogD. This might lead to unnecessary remote reads as scan ranges
are scheduled to random hosts.

This patch properly translates the host index to the coordinators host
list, so block locations remain consistent.

Testing:
 * tested manually on a 6-node cluster, and verified that the file
   locations are consistent with HDFS
 * added unit test to LocalCatalogTest

Change-Id: I253b505846e1cf4d1be445c0d06b2552dc4ba1f8
Reviewed-on: http://gerrit.cloudera.org:8080/18041
Reviewed-by: Qifan Chen <qchen@cloudera.com>
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>