IMPALA-11050: Skip filemetadata reloading in processing AlterPartition event from event processor
HdfsPartition in catalogD is a collection of files and each file is
represented by a FileDescriptor. Currently when processing an
AlterPartitionEvent, EventProcessor reloads partition metadata along
with its file metadata. File metadata reloading is an expensive
operation as it involves listing files in the underlying filesystem.

This patch makes an improvement in alter partition event processing
for external tables. The proposed solution is to compare storage
descriptor of HMS partition object received from AlterPartition event
with sd of already cached HdfsPartition. Reload file metadata if any
of the following conditions hold true:

* HdfsPartition's sd does not match sd of HMS partition object
* Partition does not exist in catalogd cache but is present in HMS
* Cache directive is present for HdfsPartition. File metadata reloading
  is required in this case to fetch latest file blocks info.

Testing:
* Added new unit tests

Change-Id: I238b169f7f1122c62cbeb1434dbb675629a1e5f2
Reviewed-on: http://gerrit.cloudera.org:8080/18083
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>